requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
google-generativeai
python-dotenv
typing-extensions

